+++ Editor/UI/RawImageEditor.cs
@@ -38,6 +38,7 @@ namespace UnityEditor.UI
 
             AppearanceControlsGUI();
             RaycastControlsGUI();
+            MaskableControlsGUI();
             EditorGUILayout.PropertyField(m_UVRect, m_UVRectContent);
             SetShowNativeSize(false);
             NativeSizeButtonGUI();