+++ Runtime/UI/Core/Culling/IClipRegion.cs
@@ -48,5 +48,13 @@ namespace UnityEngine.UI
         /// <param name="value">The Rectangle for the clipping</param>
         /// <param name="validRect">Is the rect valid.</param>
         void SetClipRect(Rect value, bool validRect);
+
+        /// <summary>
+        /// Set the clip softness for the IClippable.
+        ///
+        /// The softness is a linear alpha falloff over clipSoftness pixels.
+        /// </summary>
+        /// <param name="clipSoftness">The number of pixels to apply the softness to </param>
+        void SetClipSoftness(Vector2 clipSoftness);
     }
 }