+++ Runtime/UI/Core/Dropdown.cs
@@ -433,9 +433,6 @@ namespace UnityEngine.UI
                 return;
             #endif
 
-            m_AlphaTweenRunner = new TweenRunner<FloatTween>();
-            m_AlphaTweenRunner.Init(this);
-
             if (m_CaptionImage)
                 m_CaptionImage.enabled = (m_CaptionImage.sprite != null);
 
@@ -445,6 +442,8 @@ namespace UnityEngine.UI
 
         protected override void Start()
         {
+            m_AlphaTweenRunner = new TweenRunner<FloatTween>();
+            m_AlphaTweenRunner.Init(this);
             base.Start();
 
             RefreshShownValue();
@@ -554,7 +553,8 @@ namespace UnityEngine.UI
         /// </example>
         public void AddOptions(List<string> options)
         {
-            for (int i = 0; i < options.Count; i++)
+            var optionsCount = options.Count;
+            for (int i = 0; i < optionsCount; i++)
                 this.options.Add(new OptionData(options[i]));
             RefreshShownValue();
         }
@@ -568,7 +568,8 @@ namespace UnityEngine.UI
         /// </remarks>
         public void AddOptions(List<Sprite> options)
         {
-            for (int i = 0; i < options.Count; i++)
+            var optionsCount = options.Count;
+            for (int i = 0; i < optionsCount; i++)
                 this.options.Add(new OptionData(options[i]));
             RefreshShownValue();
         }
@@ -731,10 +732,11 @@ namespace UnityEngine.UI
                 return;
 
             // case 1064466 rootCanvas should be last element returned by GetComponentsInParent()
-            Canvas rootCanvas = list[list.Count - 1];
-            for (int i = 0; i < list.Count; i++)
+            var listCount = list.Count;
+            Canvas rootCanvas = list[listCount - 1];
+            for (int i = 0; i < listCount; i++)
             {
-                if (list[i].isRootCanvas)
+                if (list[i].isRootCanvas || list[i].overrideSorting)
                 {
                     rootCanvas = list[i];
                     break;
@@ -785,7 +787,8 @@ namespace UnityEngine.UI
             m_Items.Clear();
 
             Toggle prev = null;
-            for (int i = 0; i < options.Count; ++i)
+            var optionsCount = options.Count;
+            for (int i = 0; i < optionsCount; ++i)
             {
                 OptionData data = options[i];
                 DropdownItem item = AddItem(data, value == i, itemTemplate, m_Items);
@@ -853,12 +856,13 @@ namespace UnityEngine.UI
                     RectTransformUtility.FlipLayoutOnAxis(dropdownRectTransform, axis, false, false);
             }
 
-            for (int i = 0; i < m_Items.Count; i++)
+            var itemsCount = m_Items.Count;
+            for (int i = 0; i < itemsCount; i++)
             {
                 RectTransform itemRect = m_Items[i].rectTransform;
                 itemRect.anchorMin = new Vector2(itemRect.anchorMin.x, 0);
                 itemRect.anchorMax = new Vector2(itemRect.anchorMax.x, 0);
-                itemRect.anchoredPosition = new Vector2(itemRect.anchoredPosition.x, offsetMin.y + itemSize.y * (m_Items.Count - 1 - i) + itemSize.y * itemRect.pivot.y);
+                itemRect.anchoredPosition = new Vector2(itemRect.anchoredPosition.x, offsetMin.y + itemSize.y * (itemsCount - 1 - i) + itemSize.y * itemRect.pivot.y);
                 itemRect.sizeDelta = new Vector2(itemRect.sizeDelta.x, itemSize.y);
             }
 
@@ -1084,7 +1088,8 @@ namespace UnityEngine.UI
 
         private void ImmediateDestroyDropdownList()
         {
-            for (int i = 0; i < m_Items.Count; i++)
+            var itemsCount = m_Items.Count;
+            for (int i = 0; i < itemsCount; i++)
             {
                 if (m_Items[i] != null)
                     DestroyItem(m_Items[i]);