+++ Runtime/UI/Core/MultipleDisplayUtilities.cs
@@ -45,8 +45,8 @@ namespace UnityEngine.UI
                 // it is relative to the system resolution in order to correctly determine the position on other displays.
 
                 // Correct the y position if we are outside the main display.
-                if (position.y < 0 || position.y > Display.main.renderingHeight ||
-                    position.x < 0 || position.x > Display.main.renderingWidth)
+                if ((position.y < 0 || position.y > Display.main.renderingHeight ||
+                     position.x < 0 || position.x > Display.main.renderingWidth) && (Screen.fullScreenMode != FullScreenMode.Windowed))
                 {
                     position.y += Display.main.systemHeight - Display.main.renderingHeight;
                 }