+++ Runtime/UI/Core/ScrollRect.cs
@@ -497,7 +497,10 @@ namespace UnityEngine.UI
         private RectTransform m_HorizontalScrollbarRect;
         private RectTransform m_VerticalScrollbarRect;
 
+        // field is never assigned warning
+        #pragma warning disable 649
         private DrivenRectTransformTracker m_Tracker;
+        #pragma warning restore 649
 
         protected ScrollRect()
         {}
@@ -777,6 +780,9 @@ namespace UnityEngine.UI
             if (!this.CompareEventSystemID(eventData))
                 // exit method execution
                 return;
+            
+            if (!m_Dragging)
+                return;
 
             if (eventData.button != PointerEventData.InputButton.Left)
                 return;
@@ -1055,13 +1061,13 @@ namespace UnityEngine.UI
             // Where the position of the lower left corner of the content bounds should be, in the space of the view.
             float contentBoundsMinPosition = m_ViewBounds.min[axis] - value * hiddenLength;
             // The new content localPosition, in the space of the view.
-            float newLocalPosition = m_Content.localPosition[axis] + contentBoundsMinPosition - m_ContentBounds.min[axis];
+            float newAnchoredPosition = m_Content.anchoredPosition[axis] + contentBoundsMinPosition - m_ContentBounds.min[axis];
 
-            Vector3 localPosition = m_Content.localPosition;
-            if (Mathf.Abs(localPosition[axis] - newLocalPosition) > 0.01f)
+            Vector3 anchoredPosition = m_Content.anchoredPosition;
+            if (Mathf.Abs(anchoredPosition[axis] - newAnchoredPosition) > 0.01f)
             {
-                localPosition[axis] = newLocalPosition;
-                m_Content.localPosition = localPosition;
+                anchoredPosition[axis] = newAnchoredPosition;
+                m_Content.anchoredPosition = anchoredPosition;
                 m_Velocity[axis] = 0;
                 UpdateBounds();
             }
@@ -1143,6 +1149,7 @@ namespace UnityEngine.UI
         public virtual void SetLayoutHorizontal()
         {
             m_Tracker.Clear();
+            UpdateCachedData();
 
             if (m_HSliderExpand || m_VSliderExpand)
             {