+++ Tests/Runtime/MaskClipping/RectMask2DClipping.cs
@@ -10,11 +10,12 @@ using UnityEditor;
  This test checks that a maskableGraphic within a RectMask2D will be properly clipped.
  Test for case (1013182 - [RectMask2D] Child gameObject is masked if there are less than 2 corners of GO that matches RectMask's x position)
 */
-namespace Graphics
+namespace UnityEngine.UI.Tests
 {
     public class RectMask2DClipping : IPrebuildSetup
     {
         GameObject m_PrefabRoot;
+        GameObject m_CameraGO;
 
         const string kPrefabPath = "Assets/Resources/Mask2DRectCullingPrefab.prefab";
 
@@ -26,14 +27,14 @@ namespace Graphics
             rootCanvasGO.GetComponent<Canvas>().renderMode = RenderMode.ScreenSpaceOverlay;
             rootCanvasGO.transform.SetParent(rootGO.transform);
 
-            var maskGO = new GameObject("Mask", typeof(RectMask2D), typeof(RectTransform));
+            var maskGO = new GameObject("Mask", typeof(RectTransform), typeof(RectMask2D));
             var maskTransform = maskGO.GetComponent<RectTransform>();
             maskTransform.SetParent(rootCanvasGO.transform);
             maskTransform.localPosition = Vector3.zero;
             maskTransform.sizeDelta = new Vector2(200, 200);
             maskTransform.localScale = Vector3.one;
 
-            var imageGO = new GameObject("Image", typeof(Image), typeof(RectTransform));
+            var imageGO = new GameObject("Image", typeof(RectTransform), typeof(ImageHook));
             var imageTransform = imageGO.GetComponent<RectTransform>();
             imageTransform.SetParent(maskTransform);
             imageTransform.localPosition = new Vector3(-125, 0, 0);
@@ -53,7 +54,7 @@ namespace Graphics
         public void TestSetup()
         {
             m_PrefabRoot = Object.Instantiate(Resources.Load("Mask2DRectCullingPrefab")) as GameObject;
-            new GameObject("Camera", typeof(Camera));
+            m_CameraGO = new GameObject("Camera", typeof(Camera));
         }
 
         [UnityTest]
@@ -73,10 +74,28 @@ namespace Graphics
             Assert.IsFalse(cull);
         }
 
+        [Test]
+        public void Mask2DRect_NonZeroPaddingMasksProperly()
+        {
+            var mask = m_PrefabRoot.GetComponentInChildren<RectMask2D>();
+            mask.padding = new Vector4(10, 10, 10, 10);
+
+            Canvas.ForceUpdateCanvases();
+
+            var image = m_PrefabRoot.GetComponentInChildren<ImageHook>();
+
+            // The mask rect is 200, and padding of 10 all around makes it so its 180 in size.
+            Assert.AreEqual(180, image.cachedClipRect.height);
+            Assert.AreEqual(180, image.cachedClipRect.width);
+
+            Assert.AreEqual(-90f, image.cachedClipRect.x);
+        }
+
         [TearDown]
         public void TearDown()
         {
             Object.DestroyImmediate(m_PrefabRoot);
+            GameObject.DestroyImmediate(m_CameraGO);
         }
 
         [OneTimeTearDown]