+++ Tests/Runtime/Util/ImageHook.cs
@@ -11,14 +11,13 @@ namespace UnityEngine.UI.Tests
         public bool isGeometryUpdated;
         public bool isLayoutRebuild;
         public bool isMaterialRebuilt;
-        public bool isCacheUsed;
+        public Rect cachedClipRect;
 
         public void ResetTest()
         {
             isGeometryUpdated = false;
             isLayoutRebuild = false;
             isMaterialRebuilt = false;
-            isCacheUsed = false;
         }
 
         public override void SetLayoutDirty()
@@ -37,8 +36,15 @@ namespace UnityEngine.UI.Tests
         {
             base.UpdateGeometry();
             isGeometryUpdated = true;
-            FieldInfo fieldInfo = typeof(Image).GetField("m_UseCache", BindingFlags.Instance | BindingFlags.NonPublic);
-            isCacheUsed = (bool)fieldInfo.GetValue(gameObject.GetComponent<ImageHook>());
+        }
+
+        public override void SetClipRect(Rect clipRect, bool validRect)
+        {
+            cachedClipRect = clipRect;
+            if (validRect)
+                canvasRenderer.EnableRectClipping(clipRect);
+            else
+                canvasRenderer.DisableRectClipping();
         }
     }
 }