+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2020-04-01
+This is the second release of MCU (Multiple Concurrent Unity-UI) as a package  
+This version is based on unity-ui [1.0.1]
+
 ## [1.0.1] - 2019-13-12
 This is the first release of MCU (Multiple Concurrent Unity-UI) as a package  
 This version is based on unity-ui [1.0.0]